[AOS] rename: Component 네이밍 변경
