+++ dodamduck_aos/app/src/main/java/org/chosun/dodamduck/ui/component/lazy_components/ToyLazy.kt
@@ -1,4 +1,4 @@
-package org.chosun.dodamduck.ui.component
+package org.chosun.dodamduck.ui.component.lazy_components
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Arrangement
@@ -24,11 +24,11 @@ import androidx.compose.material3.OutlinedButton
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import org.chosun.dodamduck.R
 import org.chosun.dodamduck.model.dto.ToyInfo
+import org.chosun.dodamduck.ui.component.EllipsisText
 
 @Composable
 @Preview(showBackground = true)
